Initial commit. Working exercise #1.
Worked out name issue: don't use apostrophes in project name, such as "Mike's Anything". Since @string/app_name doesn't match project directory name, it turns out that a change is needed in the reference to the MainActivity in the property 'tools:context' of the master RelativeLayout in activity_main.xml. It takes the name of the res string by default and must be adjusted to match the folder name.